move structs to base.go
